Finalização Do método de compra
Foi feito o método de compra realizar o request do api criado na 6 questão.

+++ ApiPagamento/src/Api.Domain/Dtos/PagamentoDtoCreate.cs
@@ -4,7 +4,7 @@ namespace Api.Domain.Dtos
 {
     public class PagamentoDtoCreate
     {
-        public int valor { get; set; }
+        public decimal valor { get; set; }
         public cartao Cartao { get; set; }
     }
 }
