+++ ApiPagamento/src/Api.Domain/Models/PagamentoModel.cs
@@ -4,7 +4,7 @@ namespace Api.Domain.Models
 {
     public class PagamentoModel
     {
-        public int valor { get; set; }
+        public decimal valor { get; set; }
         public cartao Cartao { get; set; }
     }
 }
