+++ ApiProdutoCompra/src/Api.Domain/Dtos/Products/RequestExternoDto.cs
@@ -2,7 +2,7 @@ namespace Api.Domain.Dtos.Products
 {
     public class RequestExternoDto
     {
-        public int valor { get; set; }
+        public decimal valor { get; set; }
         public string estado { get; set; }
     }
 }
